backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-socketio==5.10.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis and caching
redis==5.0.1
hiredis==2.2.3

# Background tasks
celery==5.3.4
kombu==5.3.4

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=42.0.0,<47.0.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
